samples/04-AGENTIC-RAG-ROUTERS/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b6
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-cosmos==4.9.0
azure-identity==1.20.0
azure-search-documents==11.6.0b9
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fqdn==1.5.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.7
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
instructor==1.7.2
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.3
openai==1.63.2
openapi-core==0.19.4
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pathable==0.4.4
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
prance==23.6.21.0
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pure_eval==0.2.3
pybars4==0.9.13
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMeta3==0.5.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
ruamel.yaml==0.18.10
scipy==1.15.2
semantic-kernel==1.21.3
Send2Trash==1.8.3
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.29.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
